requirements.txt
pymupdf==1.25.5
pinecone==6.0.2
python-dotenv==1.1.0
logging==0.4.9.6
transformers==4.51.3
torch==2.7.0
einops==0.8.1